+++ database/migrations/2023_05_16_104622_create_finance_payments_table.php
@@ -16,9 +16,9 @@ class CreateFinancePaymentsTable extends Migration
         Schema::create('finance_payments', function (Blueprint $table) {
             $table->id();
             $table->string('currency')->nullable(true);
-            $table->float('tourist_course', 11, 2)->unsigned()->nullable(true);
-            $table->float('tour_price', 11, 2)->unsigned()->nullable(true);
-            $table->float('comission_price', 11, 2)->unsigned()->nullable(true);
+            $table->decimal('tourist_course', 11, 2)->unsigned()->nullable(true);
+            $table->decimal('tour_price', 11, 2)->unsigned()->nullable(true);
+            $table->decimal('comission_price', 11, 2)->unsigned()->nullable(true);
 
             $table->foreignId('claim_id')->onDelete('cascade')->constrained('claims');
             $table->timestamps();
